Try 8-bit (still off)

+++ examples/zigbee/light_sample/HA_on_off_light/main/light_driver.c
@@ -57,7 +57,7 @@ void light_driver_set_color(uint32_t r, uint32_t g, uint32_t b)
   ESP_ERROR_CHECK(led_strip_refresh(s_led_strip));
 }
 
-void light_driver_set_hue_and_saturation(uint16_t hue, uint16_t sat)
+void light_driver_set_hue_and_saturation(uint8_t hue, uint8_t sat)
 {
   // Placeholder: set greenish for now.
   s_red = sat;
@@ -67,7 +67,7 @@ void light_driver_set_hue_and_saturation(uint16_t hue, uint16_t sat)
   ESP_ERROR_CHECK(led_strip_refresh(s_led_strip));
 }
 
-void light_driver_set_brightness(uint16_t brightness)
+void light_driver_set_brightness(uint8_t brightness)
 {
   // Placeholder: set brightness but only for white.
   s_red = brightness;
